fix(db): remove unwanted migrations
Author: Rudy Savini - @SavGRY
