+++ backend/alembic/versions/2025_03_22_1750-3212d54ca434_.py
@@ -19,81 +19,6 @@ branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
 
-# def upgrade() -> None:
-#     # ### commands auto generated by Alembic - please adjust! ###
-#     op.drop_table('user')
-#     op.drop_table('client')
-#     op.drop_table('document')
-#     op.drop_table('incasso')
-#     # ### end Alembic commands ###
-#     pass
-#
-#
-# def downgrade() -> None:
-#     # ### commands auto generated by Alembic - please adjust! ###
-#     op.create_table(
-#         "incasso",
-#         sa.Column("id", sa.INTEGER(), autoincrement=True, nullable=False),
-#         sa.Column(
-#             "type_of_payment",
-#             postgresql.ENUM("check", "cash", name="typeofincasso"),
-#             autoincrement=False,
-#             nullable=False,
-#         ),
-#         sa.Column(
-#             "amount",
-#             sa.DOUBLE_PRECISION(precision=53),
-#             autoincrement=False,
-#             nullable=False,
-#         ),
-#         sa.Column("client_code", sa.INTEGER(), autoincrement=False, nullable=False),
-#         sa.ForeignKeyConstraint(
-#             ["client_code"],
-#             ["client.code"],
-#             name="incasso_client_code_fkey",
-#             ondelete="CASCADE",
-#         ),
-#         sa.PrimaryKeyConstraint("id", name="incasso_pkey"),
-#     )
-#     op.create_table(
-#         "document",
-#         sa.Column("id", sa.INTEGER(), autoincrement=True, nullable=False),
-#         sa.Column(
-#             "creation_date", postgresql.TIMESTAMP(), autoincrement=False, nullable=False
-#         ),
-#         sa.Column("user_id", sa.INTEGER(), autoincrement=False, nullable=False),
-#         sa.ForeignKeyConstraint(
-#             ["user_id"], ["user.id"], name="document_user_id_fkey", ondelete="CASCADE"
-#         ),
-#         sa.PrimaryKeyConstraint("id", name="document_pkey"),
-#     )
-#     op.create_table(
-#         "client",
-#         sa.Column("code", sa.INTEGER(), autoincrement=True, nullable=False),
-#         sa.Column("address", sa.VARCHAR(), autoincrement=False, nullable=True),
-#         sa.Column("name", sa.VARCHAR(), autoincrement=False, nullable=False),
-#         sa.Column("city", sa.VARCHAR(), autoincrement=False, nullable=False),
-#         sa.Column(
-#             "province", sa.VARCHAR(length=2), autoincrement=False, nullable=False
-#         ),
-#         sa.PrimaryKeyConstraint("code", name="client_pkey"),
-#     )
-#     op.create_table(
-#         "user",
-#         sa.Column("id", sa.INTEGER(), autoincrement=True, nullable=False),
-#         sa.Column("email", sa.VARCHAR(), autoincrement=False, nullable=False),
-#         sa.Column("first_name", sa.VARCHAR(), autoincrement=False, nullable=True),
-#         sa.Column("last_name", sa.VARCHAR(), autoincrement=False, nullable=True),
-#         sa.Column("is_superuser", sa.BOOLEAN(), autoincrement=False, nullable=False),
-#         sa.Column("is_active", sa.BOOLEAN(), autoincrement=False, nullable=False),
-#         sa.Column("password", sa.VARCHAR(), autoincrement=False, nullable=False),
-#         sa.Column("token", sa.VARCHAR(), autoincrement=False, nullable=True),
-#         sa.PrimaryKeyConstraint("id", name="user_pkey"),
-#         sa.UniqueConstraint("email", name="user_email_key"),
-#     )
-#     # ### end Alembic commands ###
-
-
 def upgrade() -> None:
     op.create_table(
         "user",
