--- backend/alembic/versions/2025_03_22_1823-096ba10f2137_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 096ba10f2137
-Revises: 13c8cf8c6455
-Create Date: 2025-03-22 18:23:54.238246
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "096ba10f2137"
-down_revision: Union[str, None] = "13c8cf8c6455"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
