--- backend/alembic/versions/2025_03_23_1011-137192676c1a_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 137192676c1a
-Revises: 7a326e0dfc03
-Create Date: 2025-03-23 10:11:44.647727
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "137192676c1a"
-down_revision: Union[str, None] = "7a326e0dfc03"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
