--- backend/alembic/versions/2025_03_23_1224-54404a23aa47_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 54404a23aa47
-Revises: 7899e9020213
-Create Date: 2025-03-23 12:24:18.859241
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "54404a23aa47"
-down_revision: Union[str, None] = "7899e9020213"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
