--- backend/alembic/versions/2025_03_23_1228-805950a1573f_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 805950a1573f
-Revises: 54404a23aa47
-Create Date: 2025-03-23 12:28:44.995787
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "805950a1573f"
-down_revision: Union[str, None] = "54404a23aa47"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
