--- backend/alembic/versions/2025_03_23_1229-6fda9a5d6498_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 6fda9a5d6498
-Revises: 805950a1573f
-Create Date: 2025-03-23 12:29:59.667736
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "6fda9a5d6498"
-down_revision: Union[str, None] = "805950a1573f"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
