--- backend/alembic/versions/2025_03_23_1234-782479396279_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 782479396279
-Revises: 6fda9a5d6498
-Create Date: 2025-03-23 12:34:56.552937
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "782479396279"
-down_revision: Union[str, None] = "6fda9a5d6498"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
