--- backend/alembic/versions/2025_03_23_1236-a8aabf608bd8_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: a8aabf608bd8
-Revises: 782479396279
-Create Date: 2025-03-23 12:36:10.546841
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "a8aabf608bd8"
-down_revision: Union[str, None] = "782479396279"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
