--- backend/alembic/versions/2025_03_23_1237-ec5b23ddc777_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: ec5b23ddc777
-Revises: a8aabf608bd8
-Create Date: 2025-03-23 12:37:55.478940
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "ec5b23ddc777"
-down_revision: Union[str, None] = "a8aabf608bd8"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
