--- backend/alembic/versions/2025_03_23_1238-6e3ff3497acd_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 6e3ff3497acd
-Revises: ec5b23ddc777
-Create Date: 2025-03-23 12:38:27.140944
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "6e3ff3497acd"
-down_revision: Union[str, None] = "ec5b23ddc777"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
