--- backend/alembic/versions/2025_03_23_1437-00084e213aa4_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 00084e213aa4
-Revises: 00bdb0065867
-Create Date: 2025-03-23 14:37:13.173538
-
-"""
-
-from typing import Sequence, Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "00084e213aa4"
-down_revision: Union[str, None] = "00bdb0065867"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
