+++ backend/alembic/versions/2025_04_07_1607-ee02197ee3dd_.py
@@ -1,0 +1,40 @@
+"""empty message
+
+Revision ID: ee02197ee3dd
+Revises: 00bdb0065867
+Create Date: 2025-04-07 16:07:55.918140
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = 'ee02197ee3dd'
+down_revision: Union[str, None] = '00bdb0065867'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('incasso', 'type_of_payment')
+    op.add_column('payment', sa.Column('id_payment', sa.Integer(), nullable=False))
+    op.add_column('payment', sa.Column('type_of_payment', sa.Enum('check', 'cash', name='typeofpayment'), nullable=False))
+    op.add_column('payment', sa.Column('amount', sa.Float(), nullable=False))
+    op.add_column('payment', sa.Column('client_code', sa.Integer(), nullable=False))
+    op.create_foreign_key(None, 'payment', 'client', ['client_code'], ['code'])
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'payment', type_='foreignkey')
+    op.drop_column('payment', 'client_code')
+    op.drop_column('payment', 'amount')
+    op.drop_column('payment', 'type_of_payment')
+    op.drop_column('payment', 'id_payment')
+    op.add_column('incasso', sa.Column('type_of_payment', postgresql.ENUM('check', 'cash', name='typeofincasso'), autoincrement=False, nullable=False))
+    # ### end Alembic commands ###
